requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
boto3==1.34.102
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-magic==0.4.27
slowapi==0.1.9
itsdangerous==2.1.2
python-json-logger==2.0.7
Pillow==10.3.0
bcrypt==4.1.3
prometheus_client==0.20.1 # Real metrics
starlette==0.37.2        # Core component for middleware